+++ device/util.go
@@ -1,137 +1,14 @@
-package main
+package device
 
 import (
+	"GADS/util"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
-	"os"
 	"strings"
 )
-
-//=================//
-//=====STRUCTS=====//
-
-type JsonErrorResponse struct {
-	EventName    string `json:"event"`
-	ErrorMessage string `json:"error_message"`
-}
-
-type JsonResponse struct {
-	Message string `json:"message"`
-}
-
-//=======================//
-//=====API FUNCTIONS=====//
-
-// Write to a ResponseWriter an event and message with a response code
-func JSONError(w http.ResponseWriter, event string, error_string string, code int) {
-	var errorMessage = JsonErrorResponse{
-		EventName:    event,
-		ErrorMessage: error_string}
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	w.Header().Set("X-Content-Type-Options", "nosniff")
-	w.WriteHeader(code)
-	json.NewEncoder(w).Encode(errorMessage)
-}
-
-// Write to a ResponseWriter an event and message with a response code
-func SimpleJSONResponse(w http.ResponseWriter, response_message string, code int) {
-	var message = JsonResponse{
-		Message: response_message,
-	}
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	w.Header().Set("X-Content-Type-Options", "nosniff")
-	w.WriteHeader(code)
-	json.NewEncoder(w).Encode(message)
-}
-
-// Upload application to the /apps folder to make available for Appium
-func UploadApp(w http.ResponseWriter, r *http.Request) {
-	file, header, err := r.FormFile("file")
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	defer file.Close()
-
-	// Create the apps folder if it doesn't
-	// already exist
-	err = os.MkdirAll("./apps", os.ModePerm)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	// Create a new file in the uploads directory
-	dst, err := os.Create("./apps/" + header.Filename)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	defer dst.Close()
-
-	// Copy the uploaded file to the filesystem
-	// at the specified destination
-	_, err = io.Copy(dst, file)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	fmt.Fprintf(w, "Uploaded '"+header.Filename+"' to the ./apps folder.")
-}
-
-//=======================//
-//=====FUNCTIONS=====//
-
-// Convert interface into JSON string
-func ConvertToJSONString(data interface{}) string {
-	b, err := json.MarshalIndent(data, "", "  ")
-	if err != nil {
-		fmt.Println(err)
-		return ""
-	}
-	return string(b)
-}
-
-// Unmarshal provided JSON string into a struct
-func UnmarshalJSONString(jsonString string, v interface{}) error {
-	bs := []byte(jsonString)
-
-	err := json.Unmarshal(bs, v)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Get a ConfigJsonData pointer with the current configuration from config.json
-func GetConfigJsonData() *ConfigJsonData {
-	var data ConfigJsonData
-	jsonFile, err := os.Open("./config.json")
-	if err != nil {
-		panic(err)
-	}
-	defer jsonFile.Close()
-
-	bs, err := ioutil.ReadAll(jsonFile)
-	if err != nil {
-		panic(err)
-	}
-
-	err = json.Unmarshal(bs, &data)
-	if err != nil {
-		panic(err)
-	}
-
-	return &data
-}
 
 func CheckWDASession(wdaURL string) (string, error) {
 	response, err := http.Get("http://" + wdaURL + "/status")
@@ -225,7 +102,7 @@ func checkAppiumSession(appiumURL string) (string, error) {
 	responseBody, _ := io.ReadAll(response.Body)
 
 	var responseJson AppiumGetSessionsResponse
-	err = UnmarshalJSONString(string(responseBody), &responseJson)
+	err = util.UnmarshalJSONString(string(responseBody), &responseJson)
 	if err != nil {
 		return "", err
 	}
@@ -273,10 +150,20 @@ func createAppiumSession(appiumURL string) (string, error) {
 
 	responseBody, _ := io.ReadAll(response.Body)
 	var responseJson AppiumCreateSessionResponse
-	err = UnmarshalJSONString(string(responseBody), &responseJson)
+	err = util.UnmarshalJSONString(string(responseBody), &responseJson)
 	if err != nil {
 		return "", err
 	}
 
 	return responseJson.Value.SessionID, nil
 }
+
+// func getDeviceByUDID(udid string) *Device {
+// 	for _, device := range util.ConfigData.Devices {
+// 		if device.UDID == udid {
+// 			return device
+// 		}
+// 	}
+
+// 	return nil
+// }
